ci(medium): Fix slop finder bugs and remove test redundancies (#6612)
