+++ tests/unit/scripts/gemini-client.test.ts
@@ -3,7 +3,53 @@ import {
   cleanJsonOutput,
   ReviewContext,
   FailedCheck,
+  JsonProcessor,
 } from '../../../scripts/gemini-client'
+
+describe('JsonProcessor', () => {
+  let processor: JsonProcessor
+
+  beforeEach(() => {
+    processor = new JsonProcessor()
+  })
+
+  it('should parse valid JSON and ensure labels field exists', () => {
+    const input = '{"reviewComment": "Looks good"}'
+    const result = processor.process(input)
+    expect(result.success).toBe(true)
+    expect(result.data).toEqual({
+      reviewComment: 'Looks good',
+      labels: [],
+    })
+  })
+
+  it('should not overwrite existing labels field', () => {
+    const input = '{"reviewComment": "Needs work", "labels": ["bug"]}'
+    const result = processor.process(input)
+    expect(result.success).toBe(true)
+    expect(result.data).toEqual({
+      reviewComment: 'Needs work',
+      labels: ['bug'],
+    })
+  })
+
+  it('should handle JSON within a markdown block', () => {
+    const input = '```json\n{"reviewComment": "Great job!"}\n```'
+    const result = processor.process(input)
+    expect(result.success).toBe(true)
+    expect(result.data).toEqual({
+      reviewComment: 'Great job!',
+      labels: [],
+    })
+  })
+
+  it('should return an error for invalid JSON', () => {
+    const input = '{"reviewComment": "Missing quote}'
+    const result = processor.process(input)
+    expect(result.success).toBe(false)
+    expect(result.data).toHaveProperty('error')
+  })
+})
 
 describe('cleanJsonOutput', () => {
   it('should remove markdown code blocks with "json" identifier', () => {
